backport overflow checks from 1.218 and 1.219: add a check that the arguments to isavail don't overflow. callers should probably check too, but checking here won't hurt. possible panic reported by tim newsham. Check for wraparound before the "commit" phase of uvm_map() and uvm_mapanon(), to prevent hitting assertions and/or corrupting data structures during that phase.

+++ usr.sbin/lpr/lpd/recvjob.c
@@ -1,3 +1,6 @@
+/*	$OpenBSD: recvjob.c,v 1.25 2009/10/27 23:59:52 deraadt Exp $	*/
+/*	$NetBSD: recvjob.c,v 1.14 2001/12/04 22:52:44 christos Exp $	*/
+
 /*
  * Copyright (c) 1983, 1993
  *	The Regents of the University of California.  All rights reserved.
@@ -11,11 +14,7 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *	This product includes software developed by the University of
- *	California, Berkeley and its contributors.
- * 4. Neither the name of the University nor the names of its contributors
+ * 3. Neither the name of the University nor the names of its contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
  *
@@ -32,21 +31,11 @@
  * SUCH DAMAGE.
  */
 
-#ifndef lint
-static char copyright[] =
-"@(#) Copyright (c) 1983, 1993\n\
-	The Regents of the University of California.  All rights reserved.\n";
-#endif /* not lint */
-
-#ifndef lint
-static char sccsid[] = "@(#)recvjob.c	8.1 (Berkeley) 6/6/93";
-#endif /* not lint */
-
 /*
  * Receive printer jobs from the network, queue them and
  * start the printer daemon.
  */
-#include <sys/param.h>
+#include <sys/types.h>
 #include <sys/mount.h>
 #include <sys/stat.h>
 
@@ -56,31 +45,35 @@ static char sccsid[] = "@(#)recvjob.c	8.1 (Berkeley) 6/6/93";
 #include <dirent.h>
 #include <syslog.h>
 #include <stdio.h>
+#include <errno.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdarg.h>
+#include <limits.h>
 #include "lp.h"
 #include "lp.local.h"
 #include "extern.h"
 #include "pathnames.h"
 
-#define ack()	(void) write(1, sp, 1);
+#define ack()	(void)write(STDOUT_FILENO, sp, 1);
 
-static char	 dfname[40];	/* data files */
+static char	 dfname[NAME_MAX];	/* data files */
 static int	 minfree;       /* keep at least minfree blocks available */
 static char	*sp = "";
-static char	 tfname[40];	/* tmp copy of cf before linking */
+static char	 tfname[NAME_MAX];	/* tmp copy of cf before linking */
 
-static int        chksize __P((int));
-static void       frecverr __P((const char *, ...));
-static int        noresponse __P((void));
-static void       rcleanup __P((int));
-static int        read_number __P((char *));
-static int        readfile __P((char *, int));
-static int        readjob __P((void));
+static int        chksize(int);
+static void       frecverr(const char *, ...)
+	__attribute__((__format__(__printf__, 1, 2)));
+static int        noresponse(void);
+static void       rcleanup(int);
+static int        read_number(char *);
+static int        readfile(char *, int);
+static int        readjob(void);
 
 
 void
-recvjob()
+recvjob(void)
 {
 	struct stat stb;
 	int status;
@@ -102,11 +95,13 @@ recvjob()
 	if (cgetstr(bp, "lo", &LO) == -1)
 		LO = DEFLOCK;
 
-	(void) close(2);			/* set up log file */
+	(void)close(2);			/* set up log file */
+	PRIV_START;
 	if (open(LF, O_WRONLY|O_APPEND, 0664) < 0) {
 		syslog(LOG_ERR, "%s: %m", LF);
-		(void) open(_PATH_DEVNULL, O_WRONLY);
+		(void)open(_PATH_DEVNULL, O_WRONLY);
 	}
+	PRIV_END;
 
 	if (chdir(SD) < 0)
 		frecverr("%s: %s: %m", printer, SD);
@@ -118,6 +113,7 @@ recvjob()
 		}
 	} else if (stat(SD, &stb) < 0)
 		frecverr("%s: %s: %m", printer, SD);
+
 	minfree = 2 * read_number("minfree");	/* scale KB to 512 blocks */
 	signal(SIGTERM, rcleanup);
 	signal(SIGPIPE, rcleanup);
@@ -128,13 +124,13 @@ recvjob()
 
 /*
  * Read printer jobs sent by lpd and copy them to the spooling directory.
- * Return the number of jobs successfully transfered.
+ * Return the number of jobs successfully transferred.
  */
 static int
-readjob()
+readjob(void)
 {
-	register int size, nfiles;
-	register char *cp;
+	int size, nfiles;
+	char *cp;
 
 	ack();
 	nfiles = 0;
@@ -144,12 +140,15 @@ readjob()
 		 */
 		cp = line;
 		do {
-			if ((size = read(1, cp, 1)) != 1) {
+			if ((size = read(STDOUT_FILENO, cp, 1)) != 1) {
 				if (size < 0)
-					frecverr("%s: Lost connection",printer);
+					frecverr("%s: Lost connection",
+					    printer);
 				return(nfiles);
 			}
-		} while (*cp++ != '\n');
+		} while (*cp++ != '\n' && (cp - line + 1) < sizeof(line));
+		if (cp - line + 1 >= sizeof(line))
+			frecverr("readjob overflow");
 		*--cp = '\0';
 		cp = line;
 		switch (*cp++) {
@@ -169,20 +168,32 @@ readjob()
 			 * something different than what gethostbyaddr()
 			 * returns
 			 */
-			strcpy(cp + 6, from);
-			strcpy(tfname, cp);
+			strlcpy(cp + 6, from, sizeof(line) + line - cp - 6);
+			if (strchr(cp, '/'))
+				frecverr("readjob: %s: illegal path name", cp);
+			strlcpy(tfname, cp, sizeof(tfname));
 			tfname[0] = 't';
 			if (!chksize(size)) {
-				(void) write(1, "\2", 1);
+				(void)write(STDOUT_FILENO, "\2", 1);
 				continue;
 			}
+			/*
+			 * XXX
+			 * We blindly believe what the remote host puts
+			 * for the path to the df file.  In general this
+			 * is OK since we don't allow paths with '/' in
+			 * them.  Still, it would be better to sanity
+			 * check the cf file sent to us and make the
+			 * df name match the cf name we used.  That way
+			 * we avoid any possible collisions.
+			 */
 			if (!readfile(tfname, size)) {
 				rcleanup(0);
 				continue;
 			}
 			if (link(tfname, cp) < 0)
-				frecverr("%s: %m", tfname);
-			(void) unlink(tfname);
+				frecverr("link %s %s: %m", tfname, cp);
+			(void)unlink(tfname);
 			tfname[0] = '\0';
 			nfiles++;
 			continue;
@@ -193,15 +204,14 @@ readjob()
 				size = size * 10 + (*cp++ - '0');
 			if (*cp++ != ' ')
 				break;
+			if (strchr(cp, '/'))
+				frecverr("readjob: %s: illegal path name", cp);
 			if (!chksize(size)) {
-				(void) write(1, "\2", 1);
+				(void)write(STDOUT_FILENO, "\2", 1);
 				continue;
 			}
-			(void) strcpy(dfname, cp);
-			if (index(dfname, '/'))
-				frecverr("readjob: %s: illegal path name",
-					dfname);
-			(void) readfile(dfname, size);
+			(void)strlcpy(dfname, cp, sizeof(dfname));
+			(void)readfile(dfname, size);
 			continue;
 		}
 		frecverr("protocol screwup: %s", line);
@@ -212,17 +222,14 @@ readjob()
  * Read files send by lpd and copy them to the spooling directory.
  */
 static int
-readfile(file, size)
-	char *file;
-	int size;
+readfile(char *file, int size)
 {
-	register char *cp;
+	char *cp;
 	char buf[BUFSIZ];
-	register int i, j, amt;
+	int i, j, amt;
 	int fd, err;
 
-	fd = open(file, O_CREAT|O_EXCL|O_WRONLY, FILMOD);
-	if (fd < 0)
+	if ((fd = open(file, O_CREAT|O_EXCL|O_WRONLY, FILMOD)) < 0)
 		frecverr("readfile: %s: illegal path name: %m", file);
 	ack();
 	err = 0;
@@ -232,7 +239,7 @@ readfile(file, size)
 		if (i + amt > size)
 			amt = size - i;
 		do {
-			j = read(1, cp, amt);
+			j = read(STDOUT_FILENO, cp, amt);
 			if (j <= 0)
 				frecverr("Lost connection");
 			amt -= j;
@@ -246,11 +253,12 @@ readfile(file, size)
 			break;
 		}
 	}
-	(void) close(fd);
+	(void)close(fd);
 	if (err)
 		frecverr("%s: write error", file);
 	if (noresponse()) {		/* file sent had bad data in it */
-		(void) unlink(file);
+		if (strchr(file, '/') == NULL)
+			(void)unlink(file);
 		return(0);
 	}
 	ack();
@@ -258,11 +266,11 @@ readfile(file, size)
 }
 
 static int
-noresponse()
+noresponse(void)
 {
 	char resp;
 
-	if (read(1, &resp, 1) != 1)
+	if (read(STDOUT_FILENO, &resp, 1) != 1)
 		frecverr("Lost connection");
 	if (resp == '\0')
 		return(0);
@@ -274,12 +282,13 @@ noresponse()
  * 1 == OK, 0 == Not OK.
  */
 static int
-chksize(size)
-	int size;
+chksize(int size)
 {
-	int spacefree;
+	int64_t spacefree;
 	struct statfs sfb;
 
+	if (size <= 0)
+		return (0);
 	if (statfs(".", &sfb) < 0) {
 		syslog(LOG_ERR, "%s: %m", "statfs(\".\")");
 		return (1);
@@ -292,15 +301,14 @@ chksize(size)
 }
 
 static int
-read_number(fn)
-	char *fn;
+read_number(char *fn)
 {
 	char lin[80];
-	register FILE *fp;
+	FILE *fp;
 
 	if ((fp = fopen(fn, "r")) == NULL)
 		return (0);
-	if (fgets(lin, 80, fp) == NULL) {
+	if (fgets(lin, sizeof(lin), fp) == NULL) {
 		fclose(fp);
 		return (0);
 	}
@@ -309,46 +317,35 @@ read_number(fn)
 }
 
 /*
- * Remove all the files associated with the current job being transfered.
+ * Remove all the files associated with the current job being transferred.
  */
 static void
-rcleanup(signo)
-	int signo;
+rcleanup(int signo)
 {
-	if (tfname[0])
-		(void) unlink(tfname);
-	if (dfname[0])
+	int save_errno = errno;
+
+	if (tfname[0] && strchr(tfname, '/') == NULL)
+		(void)unlink(tfname);
+	if (dfname[0] && strchr(dfname, '/') == NULL) {
 		do {
 			do
-				(void) unlink(dfname);
-			while (dfname[2]-- != 'A');
+				(void)unlink(dfname);
+			while (dfname[2]-- != 'A')
+				;
 			dfname[2] = 'z';
 		} while (dfname[0]-- != 'd');
+	}
 	dfname[0] = '\0';
+	errno = save_errno;
 }
 
-#if __STDC__
-#include <stdarg.h>
-#else
-#include <varargs.h>
-#endif
-
 static void
-#if __STDC__
 frecverr(const char *msg, ...)
-#else
-frecverr(msg, va_alist)
-	char *msg;
-        va_dcl
-#endif
 {
 	extern char fromb[];
 	va_list ap;
-#if __STDC__
+
 	va_start(ap, msg);
-#else
-	va_start(ap);
-#endif
 	rcleanup(0);
 	syslog(LOG_ERR, "%s", fromb);
 	vsyslog(LOG_ERR, msg, ap);
